Added test for window events
