+++ src/Makefile
@@ -13,7 +13,7 @@ FS_OBJS = fs.o syscall_handler_fs.o
 WINDOW_OBJS = window.o graphics.o syscall_handler_window.o window_manager.o
 PROCESS_OBJS = syscall_handler_process.o syscall_handler_permissions.o process.o permissions_capabilities.o
 
-BINARIES = bin/test_userproc_even.nun
+BINARIES = bin/test_userproc_even.nun bin/test_window.nun
 LIB_INCLUDE_PATH = ./include
 
 KERNEL_CCFLAGS = -Wall -c -ffreestanding -m32 -march=i386
