+++ src/bin/test_window.c
@@ -1,0 +1,27 @@
+int main();
+int _start() {
+    return main();
+
+    // invoke system call that kills the process
+}
+
+#include "syscall.h"
+
+int main() {
+    create_window(10, 10, 400, 400);
+    struct graphics_color c = {0,255,0};
+    struct graphics_color bg = {0,0,0};
+    struct event e;
+    int x = 0;
+    while (1) {
+        int res = get_event(&e);
+        if (res == 0) {
+            if (e.type == EVENT_KEYBOARD_PRESS) {
+                draw_char(x, 0, e.character, &c, &bg);
+                x += 10;
+            }
+        }
+    }
+
+    return 0;
+}
