+++ src/cmd_line.c
@@ -59,7 +59,9 @@ void cmd_line_attempt(const char *line) {
         cmd_line_help(the_rest);
     } else if (strcmp("window_test", first_word) == 0) {
         console_printf("\f");
-        window_hierarchy_test();
+        uint32_t identifier = permissions_capability_create();
+        run("/BIN/TEST_WIN.NUN", identifier);   
+        permissions_capability_delete(identifier);
     }
     /*else if () {
      *...
